+++ third-sem/LW2/helpers.h
@@ -7,6 +7,7 @@
 
 int *generate_array(int length, int min, int max, std::mt19937 &gen);
 
-int *remove_number_from_array(const int *array, int *fixed_array, int length, int number_to_remove, int index_to_remove);
+int *remove_number_from_array(const int *array, int *fixed_array, int length,
+                              int number_to_remove, int index_to_remove);
 
 #endif //LW_HELPERS_H
