+++ third-sem/LW2/main.cpp
@@ -36,13 +36,14 @@ const long array_sizes[] = {
         200000,
 };
 
-void perform_searches_and_save(int const *const array, long array_size, const std::vector<search_with_name> &search_pack,
-                               const std::string &pack_name) {
+void
+perform_searches_and_save(int const *const array, long array_size, const std::vector<search_with_name> &search_pack,
+                          const std::string &pack_name) {
     const std::vector<needle_def> needles = {
-            {array[array_size/10 + 1984],              array_size/10 + 1984,              true,  "start"},
-            {array[array_size * 4/10 + 1984],  array_size * 4/10 + 1984,  true,  "middle"},
-            {array[array_size * 9/10 - 1984], array_size * 9/10 - 1984, true,  "end"},
-            {42,                     -1,              false, "doesn't exist"},
+            {array[array_size / 10 + 1984],     array_size / 10 + 1984,     true,  "start"},
+            {array[array_size * 4 / 10 + 1984], array_size * 4 / 10 + 1984, true,  "middle"},
+            {array[array_size * 9 / 10 - 1984], array_size * 9 / 10 - 1984, true,  "end"},
+            {42,                                -1,                         false, "doesn't exist"},
     };
 
     auto output_path = (std::filesystem::path("results") / pack_name / std::to_string(array_size)).replace_extension(
@@ -66,9 +67,11 @@ void perform_searches_and_save(int const *const array, long array_size, const st
             fixed_array[needle.correct_index] = needle.needle;
         for (const auto &current_search: search_pack) {
             std::cout << "#Started " << current_search.name << " with the \"" << needle.name << "\" needle" << '\n';
-            search_result sort_result = current_search.func(fixed_array.get(), array_size, needle.needle, needle.correct_index);
+            search_result sort_result = current_search.func(fixed_array.get(), array_size, needle.needle,
+                                                            needle.correct_index);
             std::cout << "!Finished " << current_search.name
-                      << ". It took " << std::chrono::duration_cast<std::chrono::milliseconds>(sort_result.time_taken).count() << "ms"
+                      << ". It took "
+                      << std::chrono::duration_cast<std::chrono::milliseconds>(sort_result.time_taken).count() << "ms"
                       << " and " << sort_result.comparison_count << " comparisons" << '\n';
             fout << '|' << std::chrono::duration_cast<std::chrono::nanoseconds>(sort_result.time_taken).count() << '|'
                  << sort_result.comparison_count;
